refactor: 상품 컨텍스트에 해당하는 모듈을 product 패키지로 이동
